requirements.txt
streamlit>=1.49.0
gspread
google-auth
google-auth-oauthlib
google-auth-httplib2